Improve type support in inferences and criticisms (#561)
* raise more informative errors; add more type compatibilities

* update inference unit test

* make data/latent_vars type checking a utility function

* improve type support for criticism

* update unit tests for ed.{evaluate,ppc}

* split test_inference.py to test the utility functions

* update __init__.py

* fix inference type logic

* remove unnecessary floating point casts in examples

* streamline data logic

+++ examples/bayesian_linear_regression_implicitklqp.py
@@ -30,7 +30,7 @@ from tensorflow.contrib import slim
 
 def build_toy_dataset(N, w, noise_std=0.1):
   D = len(w)
-  x = np.random.randn(N, D).astype(np.float32)
+  x = np.random.randn(N, D)
   y = np.dot(x, w) + np.random.normal(0, noise_std, size=N)
   return x, y
